tpl/tplimpl: Fix context in shortcode error messages
Fixes #13279
