+++ tpl/tplimpl/embedded/templates/shortcodes/qr.html
@@ -9,7 +9,7 @@ Encodes the given text into a QR code using the specified options and renders th
 @param {string} [class] The class attribute of the img element.
 @param {string} [id] The id attribute of the img element.
 @param {string} [title] The title attribute of the img element.
-@param {string} [loading] The loading attribute of the img element, one of lazy, or eager. 
+@param {string} [loading] The loading attribute of the img element, one of lazy or eager.
 
 @returns {template.HTML}
 
@@ -48,7 +48,6 @@ Encodes the given text into a QR code using the specified options and renders th
 {{- $id := or (.Get "id") "" }}
 {{- $title := or (.Get "title") "" }}
 {{- $loading := or (.Get "loading") "" }}
-
 
 {{- /* Validate arguments. */}}
 {{- $errors := false}}
